debug: 1. 解决当特征是double时，数值型无法正常处理的问题。

+++ libs/feature/udfs.py
@@ -21,7 +21,7 @@ def to_string(v):
 def weekday(dt):
     return f'{dt:%A}'.lower()
 
-@udf(returnType=IntegerType())
+@udf(returnType=DoubleType())
 def int_default_zero(values):
     if values:
         return values
